feat: editor canvas lazy improvements
