+++ apps/codeimage/src/_MobileApp.tsx
@@ -1,21 +1,11 @@
-import {createSignal, JSXElement, lazy} from 'solid-js';
-import {useFrameState} from './state/frame';
-import {useTerminalState} from './state/terminal';
-import {Toolbar} from './components/Toolbar/Toolbar';
-import {Canvas} from './components/Scaffold/Canvas/Canvas';
-import {FrameHandler} from './components/Frame/FrameHandler';
-import {Frame} from './components/Frame/Frame';
-import {DynamicTerminal} from './components/Terminal/dynamic/DynamicTerminal';
-import {CustomEditor} from './components/CustomEditor/CustomEditor';
-import {Footer} from './components/Footer/Footer';
+import {createSignal, JSXElement, lazy, Suspense} from 'solid-js';
 
 const LazyBottomBar = lazy(() => import('./components/BottomBar/BottomBar'));
 
+const LazyEditor = lazy(() => import('./components/Editor/Editor'));
+
 export default function MobileApp(): JSXElement {
-  const [frameRef, setFrameRef] = createSignal<HTMLElement>();
   const [portalHostRef, setPortalHostRef] = createSignal<HTMLElement>();
-  const frame = useFrameState();
-  const terminal = useTerminalState();
 
   return (
     <>
@@ -30,35 +20,11 @@ export default function MobileApp(): JSXElement {
           'z-index': 10,
         }}
       />
-      <Canvas>
-        <Toolbar canvasRef={frameRef()} />
-        <FrameHandler ref={setFrameRef} onScaleChange={frame.setScale}>
-          <Frame
-            radius={0}
-            padding={frame.padding}
-            background={frame.background}
-            opacity={frame.opacity}
-            visible={frame.visible}
-          >
-            <DynamicTerminal
-              type={terminal.type}
-              readonlyTab={false}
-              tabName={terminal.tabName}
-              showTab={true}
-              shadow={terminal.shadow}
-              background={terminal.background}
-              accentVisible={terminal.accentVisible}
-              darkMode={terminal.darkMode}
-              textColor={terminal.textColor}
-              onTabChange={terminal.setTabName}
-              showHeader={terminal.showHeader}
-            >
-              <CustomEditor />
-            </DynamicTerminal>
-          </Frame>
-        </FrameHandler>
-        <Footer />
-      </Canvas>
+
+      <Suspense fallback={<></>}>
+        <LazyEditor />
+      </Suspense>
+
       <LazyBottomBar portalHostRef={portalHostRef()} />
     </>
   );
